src/requirements.txt
numpy>=1.21.0
requests>=2.26.0
pyYAML>=6.0
packaging>=21.0
pydantic>=1.8.2
